+++ transducers/package.lisp
@@ -21,6 +21,7 @@
   (:export #:cons #:snoc #:vector #:string #:hash-table
            #:count #:average #:median #:quantities
            #:any? #:all? #:anyp #:allp #:any #:all
+           #:partition
            #:first #:last
            #:fold #:max #:min #:find
            #:for #:for-each)