+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 
 - Reducer: `quantities` for counting unique occurrences of streamed items.
+- Reducer: `partition` for splitting the stream results.
 
 #### Fixed